Move lxd/util.go into its own lxd/util/ sub-package
Move all the code currently in lxd/util.go to a new lxd/util
sub-package, and split it into two files according to the
functionality (http.go for http-related utils, kernel.go for
kernel-related utils).

In detail:

- Add a new lxd/util sub-package.
- Rename lxd/util.go to lxd/util/http.go, changing its package name
  from "main" to "util".
- Move loadModule from lxd/util/http.go to lxd/util/kernel.go.
- Make all functions public.
- Update call sites.
